Added generator of boundary strings
This one isn't a generator itself. A new class (and package) has been created, BoundaryStringConfigurator, which internally uses the RandomStringGenerator, that takes the former's parameters into account. When setting the generator from the factory, it will be an InputValue generator, whose values will be obtained using the returnValues() method from the BoundaryStringConfigurator class.