Refactor TestConfigurationFilter and its usages
In short this commit
- Adds a method which creates a new filter by parsing a string
- Adds a builder to enable creating filters in fluent-style
- Makes TestConfigurationFilter object immutable after creation

In many places handling of TestConfigurationFilter looked a bit clumsy
to me: first create an object, then look at HTTP method names, then set
the filter's fields. In order to make working with filters more elegant,
the logic of creating filters is now hidden in one method, which takes
a string and returns a filter.

This change allowed to make filter objects immutable, which is almost
always a good thing: the less the state changes, the less you have
to think about it.